chore: seperate test cant using race
